requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
gunicorn
python-dotenv
groq
huggingface_hub
Pillow
requests
google-generativeai

# ── Multi-Agent RAG Dependencies ──
chromadb
sentence-transformers